Finish adding docstrings, gcp module refactoring
