+++ src/provider.rs
@@ -1,3 +1,12 @@
+//! # Trusted Execution Environment (TEE) Attestation Interface
+//!
+//! This module provides the `AttestationProvider` trait, which VM-based TEE
+//! guests can use to interact with the TEE platform within the guest
+//! environment.
+//!
+//! The trait provides a function for retrieving TEE attestation reports and
+//! launch-time measurements.
+
 use crate::error::Result;
 
 pub trait AttestationProvider {
